requirements.txt
customtkinter==5.2.0
Pillow==10.0.0
